Drew Drive Captsone 2

+++ java/tenmo-server/src/main/java/com/techelevator/tenmo/dao/UserSqlDAO.java
@@ -32,22 +32,20 @@ public class UserSqlDAO implements UserDAO {
 	public List<User> findAll() {
 		List<User> users = new ArrayList<>();
 		String sql = "select * from users";
-
 		SqlRowSet results = jdbcTemplate.queryForRowSet(sql);
-		while(results.next()) {
+		while (results.next()) {
 			User user = mapRowToUser(results);
 			users.add(user);
 		}
-
 		return users;
 	}
 
 	@Override
 	public User findByUsername(String username) throws UsernameNotFoundException {
 		for (User user : this.findAll()) {
-			if( user.getUsername().toLowerCase().equals(username.toLowerCase())) {
+			if (user.getUsername().toLowerCase().equals(username.toLowerCase())) {
 				return user;
-			}			
+			}
 		}
 		throw new UsernameNotFoundException("User " + username + " was not found.");
 	}
@@ -57,36 +55,34 @@ public class UserSqlDAO implements UserDAO {
 		boolean userCreated = false;
 		boolean accountCreated = false;
 
-		// create user
+		// CREATE USER
 		String insertUser = "insert into users (username,password_hash) values(?,?)";
 		String password_hash = new BCryptPasswordEncoder().encode(password);
 
 		GeneratedKeyHolder keyHolder = new GeneratedKeyHolder();
 		String id_column = "user_id";
 		userCreated = jdbcTemplate.update(con -> {
-			PreparedStatement ps = con.prepareStatement(insertUser, new String[]{id_column});
+			PreparedStatement ps = con.prepareStatement(insertUser, new String[] { id_column });
 			ps.setString(1, username);
-			ps.setString(2,password_hash);
+			ps.setString(2, password_hash);
 			return ps;
-		}
-		, keyHolder) == 1;
+		}, keyHolder) == 1;
 		int newUserId = (int) keyHolder.getKeys().get(id_column);
 
 		// create account
 		String insertAccount = "insert into accounts (user_id,balance) values(?,?)";
-		accountCreated = jdbcTemplate.update(insertAccount,newUserId,STARTING_BALANCE) == 1;
+		accountCreated = jdbcTemplate.update(insertAccount, newUserId, STARTING_BALANCE) == 1;
 
 		return userCreated && accountCreated;
-	}    
+	}
 
 	private User mapRowToUser(SqlRowSet rs) {
 		User user = new User();
-		user.setId(rs.getInt("user_id"));
+		user.setId(rs.getLong("user_id"));
 		user.setUsername(rs.getString("username"));
 		user.setPassword(rs.getString("password_hash"));
 		user.setActivated(true);
 		user.setAuthorities("ROLE_USER");
 		return user;
 	}
-
 }
